telegramchatbot/telegramchatbot/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bson==0.5.10
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
django-crontab==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
postgres==4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pyjokes==0.6.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
